Correccion de error al borrar en Categoria y Marca,agregado mas fotos en Producto,arreglado el modal,agregado mas productos,arreglado el sistema de borrado de Categoria y marca

+++ app/Http/Controllers/CategoriaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Categoria;
+use App\Producto;
 
 class CategoriaController extends Controller
 {
@@ -13,7 +14,7 @@ class CategoriaController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
+    {  
         $categorias = Categoria::all();
         return view('abmCategoria', compact('categorias'));
     }
@@ -38,12 +39,13 @@ class CategoriaController extends Controller
     {
       $reglas = [
         "categoria" => "min:2|unique:categorias",
-        "img"=>"nullable|image|mimes:jpeg,png,jpg|min:1|max:10000000",
+        "img"=>"nullable|image|mimes:jpeg,png,jpg|min:1|max:10000000|dimensions:max_width=200,max_height=200",
       ];
 
       $msj = [
         "min" => "El campo debe tener un minimo de :min caracteres",
-        "unique" => "No se puede agregar categorias que ya estan en la base de datos"
+        "unique" => "No se puede agregar categorias que ya estan en la base de datos",
+        "dimensions"=>"El campo :attribute tiene dimensiones de imagen inválidas tiene que ser de menos 200 de alto y ancho."
       ];
 
       $this->validate($request, $reglas, $msj);
@@ -52,13 +54,12 @@ class CategoriaController extends Controller
     if($request->file("img")){
       $file=$request->file("img");
       $imagenNombre="/img/categorias/";
-      $imagenNombre.=$request->categoria."/";
       $imagenNombre.=$request->categoria;
       $imagenNombre.=".";
       $imagenNombre.=$file->getClientOriginalExtension();
-      $request->img->move(public_path("img/categorias/$request->categoria/"),$imagenNombre);
+      $request->img->move(public_path("img/categorias/"),$imagenNombre);
     }else{
-      $imagenNombre="/img/categorias/pc.png";
+      $imagenNombre="/img/nod.png";
     }
 
       $Categoria = New Categoria;
@@ -102,21 +103,48 @@ class CategoriaController extends Controller
      */
     public function update(Request $request)
     {
+      $Categoria = Categoria::find($request->id_categoria);
+      $array=Categoria::all("categoria")->where("categoria","<>",$Categoria->categoria);
+      $arrayNombres=[];
+      foreach ($array as $key => $value) {
+        $arrayNombres[]=$value->nombre;
+      }
+      $arrayNombres=implode(",",$arrayNombres);
       $reglas = [
-        "categoria" => "min:2|unique:categorias"
+        "categoria" => "notIn:$arrayNombres|required|string|min:2",
+        "img"=>"nullable|image|mimes:jpeg,png,jpg|min:1|max:10000000|dimensions:max_width=200,max_height=200",
       ];
 
       $msj = [
         "min" => "El campo debe tener un minimo de :min caracteres",
-        "unique" => "No se puede agregar categorias que ya estan en la base de datos"
+        "unique" => "No se puede agregar categorias que ya estan en la base de datos",
+        "dimensions"=>"El campo :attribute tiene dimensiones de imagen inválidas tiene que ser de menos 200 de alto y ancho."
       ];
 
       $this->validate($request, $reglas, $msj);
 
-      $Categoria = Categoria::find($request->id_categoria);
+      
 
       $Categoria->categoria = $request->categoria;
+      $imagenNombre=$Categoria->img;
+    if($request->file("img")){
+      $file=$request->file("img");
+      $imagenNombre="/img/categorias/";
+      $imagenNombre.=$request->categoria;
+      $imagenNombre.=".";
+      $imagenNombre.=$file->getClientOriginalExtension();
+      $request->img->move(public_path("img/categorias/"),$imagenNombre);
+    }else{
+      if($Categoria->categoria!=$request->categoria){
+        $nuevo=str_replace($Categoria->categoria,$request->categoria,$Categoria->img);
+        rename(public_path($Categoria->img),public_path($nuevo));
+        $imagenNombre=$nuevo;
+      }else{
+        $imagenNombre=$Categoria->img;
+      }
 
+    }
+      $Categoria->img=$imagenNombre;
       $Categoria->save();
 
       return view("modificarCategoria", compact('Categoria'));
@@ -131,10 +159,20 @@ class CategoriaController extends Controller
     public function destroy($id)
     {
       $Categoria = Categoria::find($id);
-
-      $Categoria->delete();
-
-      return redirect('cuenta/admin/abmCategoria');
+      if(isset($Categoria)){
+        $productos=Producto::all("id_categoria")->where("id_categoria","=",$id)->first();
+        //dd($productos);
+        if(!isset($productos)){
+          $msj[1]="Categoria eliminada $Categoria->categoria";
+          $Categoria->delete();
+          $msj[0]="success";
+        }else{
+          $msj[1]="No se puede eliminar la Categoria $Categoria->categoria";
+          $msj[0]="danger";
+        }
+        return view('eliminarCategoria')->with("msj",$msj);
+      }
+      return redirect("/error");
     }
