+++ app/Http/Controllers/MarcaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Marca;
+use App\Producto;
 
 class MarcaController extends Controller
 {
@@ -116,10 +117,20 @@ class MarcaController extends Controller
     public function destroy($id)
     {
 
-        $Marca = Marca::find($id);
-
-        $Marca->delete();
-
-        return redirect('cuenta/admin/abmMarca');
+      $Marca = Marca::find($id);
+      if(isset($Marca)){
+        $productos=Producto::all("id_marca")->where("id_marca","=",$id)->first();
+        //dd($productos);
+        if(!isset($productos)){
+          $msj[1]="Marca eliminada $Marca->marca";
+          $Marca->delete();
+          $msj[0]="success";
+        }else{
+          $msj[1]="No se puede eliminar la Marca $Marca->marca";
+          $msj[0]="danger";
+        }
+        return view('eliminarMarca')->with("msj",$msj);
+      }
+      return redirect("/error");
     }
 }
